Changing hoursRemaining to be a decimal due to Mixer changes

+++ Mixer.Base/Model/TestStreams/TestStreamSettingsModel.cs
@@ -7,7 +7,7 @@ namespace Mixer.Base.Model.TestStreams
         public uint id { get; set; }
         public bool? isActive { get; set; }
         public int hoursQuota { get; set; }
-        public int hoursRemaining { get; set; }
+        public double hoursRemaining { get; set; }
         public uint hoursResetIntervalInDays { get; set; }
         public DateTimeOffset? hoursLastReset { get; set; }
     }
